Long 타입으로 선언되어있던 userId -> Integer 타입으로 변경 엔티티와 DB 구조 일치하도록 수정

+++ src/main/java/ac/su/kdt/prompttest/controller/ChatHistoryController.java
@@ -27,7 +27,7 @@ public class ChatHistoryController {
     
     @GetMapping("/user/{userId}/session/{sessionId}")
     public ResponseEntity<List<ChatHistory>> getChatHistory(
-            @PathVariable Long userId,
+            @PathVariable Integer userId,
             @PathVariable String sessionId) {
         List<ChatHistory> history = chatHistoryService.getUserChatHistory(userId, sessionId);
         return ResponseEntity.ok(history);
